# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation for Dihauti Residence",
        "description": "description",
        "contact": {
            "email": "dev.arven@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.dihauti.ru/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/forumCategories": {
            "get": {
                "tags": [
                    "ForumCategory"
                ],
                "summary": "Get all forum categories",
                "description": "Returns all forum categories",
                "operationId": "getForumCategoryList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "ForumCategory"
                ],
                "summary": "Create forum category in database",
                "operationId": "forumCategoryCreate",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/forumCategories/{slug}": {
            "get": {
                "tags": [
                    "ForumCategory"
                ],
                "summary": "Get forum category by slug",
                "description": "Returns forum category by slug",
                "operationId": "getForumCategoryBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Forum category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ForumCategory"
                ],
                "summary": "Update forum category in database",
                "operationId": "forumCategoryUpdate",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Forum category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ForumCategory"
                ],
                "summary": "Delete forum category in database",
                "operationId": "forumCategoryDelete",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Forum category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get all questions with group by category",
                "description": "Returns all questions with group by category",
                "operationId": "getQuestionsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create question in database",
                "operationId": "questionCreate",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/questions/{slug}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get question by slug",
                "description": "Returns question by slug",
                "operationId": "getQuestionBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Question slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update question in database",
                "operationId": "questionUpdate",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Question slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Question"
                ],
                "summary": "Delete question in database",
                "operationId": "questionDelete",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Question slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/forumCategories/{forumCategory}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get questions with category",
                "description": "Returns questions with category",
                "operationId": "getQuestionsListWithCategory",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Forum category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/login",
                        "refreshUrl": "/logout",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}